change edgeNudger so that we can maintain infinite scroll after resizing despite our edgenudger (our listener overridge nudging thing at the browser scrollbar edges):
"Re-attach the listener every time the size changes and run the edge check in the next animation-frame so that drei has already handled the event. Also nudge by exactly 1 px instead of a percentage."

...Like, it was apparently some business with the drei listener being detached/reattached on resize which moves it to the end of the queue, and our edgenudger always fires first... since the queue... doesn't automatically push drei back after resizing for some reason?

Oh, also remove the asinine timeouts and intervals.

Anyway, the big thing is probably just adding the removeEventListener for onResize. But there's some subtlety that we'd need to explore more with requestAnimationFrame.